requirements-test.txt
pytest ~= 8.4.1
pytest-django ~= 4.11.1
tox ~= 4.27.0
flake8 ~= 7.3.0
coverage ~= 7.9.2
